requirements.txt
requests
wget
python-slugify